Add double instead of int to work better in grade

+++ EX 2.c
@@ -7,19 +7,19 @@
     char Aluno2 [20] ;
     char Aluno3 [20] ;
     char Aluno4 [20] ;
-    int Nota1;
-    int Nota2;
-    int Nota3;
-    int Nota4;
+    double Nota1;
+    double Nota2;
+    double Nota3;
+    double Nota4;
 
      printf("Insira os nomes do aluno e suas respectivas notas nota: \n");
          printf("Nomes\n");
             scanf("%s\n%s\n%s\n%s", &Aluno1, &Aluno2,&Aluno3, &Aluno4);
          printf("Notas\n");
-            scanf("%d\n%d\n%d\n%d", &Nota1, &Nota2,&Nota3, &Nota4);
+            scanf("%lf\n%lf\n%lf\n%lf", &Nota1, &Nota2,&Nota3, &Nota4);
 
      printf("Alunos:  Notas:\n");
-      printf("%s  %d\n" "%s  %d\n" "%s  %d\n" "%s  %d\n ", Aluno1, Nota1, Aluno2,  Nota2, Aluno3, Nota3, Aluno4, Nota4 );
+      printf("%s  %lf\n" "%s  %lf\n" "%s  %lf\n" "%s  %lf\n ", Aluno1, Nota1, Aluno2,  Nota2, Aluno3, Nota3, Aluno4, Nota4 );
